+++ src/weather/dropping.rs
@@ -29,7 +29,7 @@ impl GeneralDropping {
 impl WeatherImpl for GeneralDropping {}
 
 impl EachFrameImpl for GeneralDropping {
-    fn on_frame(&mut self, rb: &mut RenderBuffer, seed: u64, frame: u8) {
+    fn on_frame(&mut self, rb: &mut RenderBuffer, seed: u64, frame: u64) {
         self.wind.on_frame(rb, seed, frame);
         self.dropping.on_frame(rb, seed, frame);
     }
@@ -70,7 +70,7 @@ impl TailDropping {
 impl WeatherImpl for TailDropping {}
 
 impl EachFrameImpl for TailDropping {
-    fn on_frame(&mut self, rb: &mut RenderBuffer, seed: u64, frame: u8) {
+    fn on_frame(&mut self, rb: &mut RenderBuffer, seed: u64, frame: u64) {
         self.wind.on_frame(rb, seed, frame);
         self.dropping.on_frame(rb, seed, frame);
         self.tail.on_frame(rb, seed, frame);
